+++ R/heat_map.R
@@ -35,6 +35,10 @@
 #' International Monetary Fund (2013). *Staff Guidance Note for Public Debt
 #' Sustainability Analysis in Market-Access Countries*. IMF Policy Paper.
 #'
+#' International Monetary Fund (2022). *Staff Guidance Note on the Sovereign
+#' Risk and Debt Sustainability Framework for Market Access Countries*.
+#' IMF Policy Paper.
+#'
 #' @export
 #' @examples
 #' hm <- dk_heat_map(